# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bemantis API",
    "description": "Plataforma de gestion de leads con integracion WhatsApp IA",
    "version": "1.0.0"
  },
  "paths": {
...
    "/leads/": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "List Leads",
        "description": "Lista leads con filtros opcionales y paginacion.",
        "operationId": "list_leads_leads__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EstadoLead"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por estado",
              "title": "Estado"
            },
            "description": "Filtrar por estado"
          },
          {
            "name": "telefono",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Buscar por telefono exacto",
              "title": "Telefono"
            },
            "description": "Buscar por telefono exacto"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadsPaginados"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Leads"
        ],
        "summary": "Create Lead",
        "description": "Crea un nuevo lead. Usado por el formulario frontend y por el flujo n8n.",
        "operationId": "create_lead_leads__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leads/{lead_id}": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "Get Lead",
        "description": "Retorna detalle completo de un lead por ID.",
        "operationId": "get_lead_leads__lead_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Leads"
        ],
        "summary": "Update Lead",
        "description": "Actualiza campos parciales de un lead existente.",
        "operationId": "update_lead_leads__lead_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Leads"
        ],
        "summary": "Delete Lead",
        "description": "Elimina un lead y en cascada todas sus interacciones.",
        "operationId": "delete_lead_leads__lead_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EstadoLead": {
        "type": "string",
        "enum": [
          "cold",
          "warm",
          "hot",
          "appointment"
        ],
        "title": "EstadoLead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeadCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "empresa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa"
          },
          "correo": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correo"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "estado": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EstadoLead"
              },
              {
                "type": "null"
              }
            ],
            "default": "cold"
          },
          "fuente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuente"
          },
          "notas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notas"
          }
        },
        "type": "object",
        "required": [
          "nombre"
        ],
        "title": "LeadCreate",
        "description": "Payload para crear un nuevo lead."
      },
      "LeadResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "empresa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa"
          },
          "correo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correo"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoLead"
          },
          "fuente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuente"
          },
          "notas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notas"
          },
          "creado_en": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creado En"
          },
          "fecha_actualizacion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Actualizacion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "empresa",
          "correo",
          "telefono",
          "estado",
          "fuente",
          "notas",
          "creado_en",
          "fecha_actualizacion"
        ],
        "title": "LeadResponse",
        "description": "Representacion completa del lead para el cliente."
      },
      "LeadUpdate": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "empresa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa"
          },
          "correo": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correo"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "estado": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EstadoLead"
              },
              {
                "type": "null"
              }
            ]
          },
          "fuente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuente"
          },
          "notas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notas"
          }
        },
        "type": "object",
        "title": "LeadUpdate",
        "description": "Campos parciales para actualizar un lead."
      },
      "LeadsPaginados": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "pagina": {
            "type": "integer",
            "title": "Pagina"
          },
          "leads": {
            "items": {
              "$ref": "#/components/schemas/LeadResponse"
            },
            "type": "array",
            "title": "Leads"
          }
        },
        "type": "object",
        "required": [
          "total",
          "pagina",
          "leads"
        ],
        "title": "LeadsPaginados",
        "description": "Respuesta paginada de leads."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}